Revert "Fix broadcast er receive data"
This reverts commit eedb247db24a0c64ef998fb40d3ea2319d44db3d.
